refactor(messaging): move attachment text extraction into messaging
